Add count of dataset samples, and cache test on disk
Need to count the samples in the dataset since that can change (sometimes I use a small part of dataset for testing). The count is used to determine the validation split. Cache test on disk instead of in memory as it is not used much and it frees up memory:)